clean up hardhat config by deleting tasks and moving each deploy process to its own script
