--- sample/deploy.ts
@@ -1,110 +1,0 @@
-// Copyright (c) 2021 Curvegrid Inc.
-
-import { SignerWithAddress } from "@nomicfoundation/hardhat-ethers/signers";
-import { HardhatRuntimeEnvironment } from "hardhat/types";
-import {
-  DeployResult,
-  DeployProxyResult,
-} from "hardhat-multibaas-plugin/lib/type-extensions";
-
-export async function deployGreeterContract(
-  signer: SignerWithAddress,
-  hre: HardhatRuntimeEnvironment,
-): Promise<DeployResult> {
-  await hre.mbDeployer.setup();
-
-  return hre.mbDeployer.deploy(
-    signer as SignerWithAddress,
-    "Greeter",
-    ["Hello, world!"],
-    {
-      addressLabel: "greeter",
-      contractVersion: "1.0",
-      contractLabel: "greeter",
-    },
-  );
-}
-
-export async function deployThenLinkGreeterContract(
-  signer: SignerWithAddress,
-  hre: HardhatRuntimeEnvironment,
-): Promise<DeployResult> {
-  // This function is a demo of how to link a contract if it is already deployed,
-  // or is deployed by the normal Hardhat deployer rather than the MultiBaas deployer.
-  // It assumes that the Greeter contract is already uploaded
-  // to MultiBaas. It will fail with a 404 if the Greeter contract is not found.
-  await hre.mbDeployer.setup();
-
-  const factory = await hre.ethers.getContractFactory("Greeter", signer);
-  const contract = await factory.deploy("Hello, world!");
-  await contract.waitForDeployment();
-
-  return hre.mbDeployer.link(
-    signer as SignerWithAddress,
-    "Greeter",
-    await contract.getAddress(),
-    {
-      addressLabel: "linked_greeter",
-      contractVersion: "1.0",
-      contractLabel: "greeter",
-    },
-  );
-}
-
-export async function deployProxiedGreeterContract(
-  signer: SignerWithAddress,
-  hre: HardhatRuntimeEnvironment,
-): Promise<DeployProxyResult> {
-  await hre.mbDeployer.setup();
-
-  return hre.mbDeployer.deployProxy(
-    signer as SignerWithAddress,
-    "ProxiedGreeter",
-    ["Hello, world!"],
-    {
-      addressLabel: "proxied_greeter",
-      contractVersion: "1.0",
-      contractLabel: "proxied_greeter",
-    },
-  );
-}
-
-export async function deployMetaCoinContract(
-  signer: SignerWithAddress,
-  hre: HardhatRuntimeEnvironment,
-): Promise<DeployResult> {
-  await hre.mbDeployer.setup();
-
-  const convertLibInstance = (
-    await hre.mbDeployer.deploy(
-      signer,
-      "ConvertLib",
-      [],
-      // we don't want to deploy multiple contracts into MultiBaas
-      {
-        contractVersion: "1.0",
-        addressLabel: "convertlib",
-        contractLabel: "convertlib",
-      },
-    )
-  ).contract;
-
-  return hre.mbDeployer.deploy(
-    {
-      signer,
-      // MetaCoin must be linked with ConvertLib before deploying
-      libraries: {
-        ConvertLib: await convertLibInstance.getAddress(),
-      },
-    },
-    "MetaCoin",
-    [],
-    // we don't want to deploy multiple contracts into MultiBaas
-    {
-      contractVersion: "1.0",
-      addressLabel: "metacoin",
-      contractLabel: "metacoin",
-      startingBlock: "0",
-    },
-  );
-}
